requirements.txt
alembic==1.4.0
Babel==2.8.0
blinker==1.4
cffi==1.14.0
Click==7.0
cryptography==2.8
cycler==0.10.0
dominate==2.5.1
Flask==1.1.1
Flask-Babel==1.0.0
Flask-Bootstrap==3.3.7.1
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-Migrate==2.5.2
Flask-Moment==0.9.0
flask-paginate==0.5.5
Flask-SQLAlchemy==2.4.1
Flask-WTF==0.14.3
itsdangerous==1.1.0
Jinja2==2.11.1
kiwisolver==1.1.0
Mako==1.1.1
MarkupSafe==1.1.1
matplotlib==3.2.1
numpy==1.18.2
pycparser==2.20
PyJWT==1.7.1
pyparsing==2.4.6
python-dateutil==2.8.1
python-dotenv==0.12.0
python-editor==1.0.4
pytz==2019.3
six==1.14.0
SQLAlchemy==1.3.13
visitor==0.1.3
Werkzeug==1.0.0
WTForms==2.2.1
